Create MOVES.C
This project has been developed in C graphics and was a part of coursework during my Masters in Computer Application. It uses graphics.h header file which is obsolete now. If you run this code on modern architecture using for example TURBOC3, it will not execute completely. It is recommended to run it using the legacy version TURBOC1.